BarsHelper.ColorMode prop splitted to two TitleBar and StatusBar related props, related changes, BarsHelperColorMode.ThemedDefault renamed to Themed, updated NuGet dependencies
