+++ UWPHelper.SampleApp/App.xaml.cs
@@ -54,7 +54,7 @@ namespace UWPHelper.SampleApp
                 await loadAppDataTask;
 
                 //BarsHelper.Current.InitializeForAutoUpdate(() => AppData.Current.Theme, AppData.Current, nameof(AppData.Theme));
-                BarsHelper.Current.ColorMode = BarsHelperColorMode.ThemedGray;
+                BarsHelper.Current.StatusBarColorMode = BarsHelperColorMode.ThemedGray;
                 await BarsHelper.Current.InitializeForCurrentViewAsync();
 
                 AppData.Current.SetTheme();
