+++ UWPHelper/UI/BarsHelperTitleBarColorsSetterThemedGray.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using Windows.UI.ViewManagement;
+using Windows.UI.Xaml;
+
+namespace UWPHelper.UI
+{
+    public sealed class BarsHelperTitleBarColorsSetterThemedGray : IBarsHelperTitleBarColorsSetter
+    {
+        public void SetTitleBarColors(ApplicationViewTitleBar titleBar, ElementTheme requestedTheme)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
