+++ UWPHelper/UI/IBarsHelperStatusBarColorsSetter.cs
@@ -1,0 +1,11 @@
+﻿using Windows.Graphics.Display;
+using Windows.UI.ViewManagement;
+using Windows.UI.Xaml;
+
+namespace UWPHelper.UI
+{
+    public interface IBarsHelperStatusBarColorsSetter
+    {
+        void SetStatusBarColors(StatusBar statusBar, ElementTheme requestedTheme, bool useDarkerStatusBarOnLandscapeOrientation, DisplayOrientations currentOrientation);
+    }
+}
