+++ UWPHelper/UI/IBarsHelperTitleBarColorsSetter.cs
@@ -1,0 +1,10 @@
+﻿using Windows.UI.ViewManagement;
+using Windows.UI.Xaml;
+
+namespace UWPHelper.UI
+{
+    public interface IBarsHelperTitleBarColorsSetter
+    {
+        void SetTitleBarColors(ApplicationViewTitleBar titleBar, ElementTheme requestedTheme);
+    }
+}
