+++ UWPHelper/project.json
@@ -1,6 +1,6 @@
 ﻿{
   "dependencies": {
-    "Microsoft.NETCore.UniversalWindowsPlatform": "5.3.0",
+    "Microsoft.NETCore.UniversalWindowsPlatform": "5.3.1",
     "Microsoft.Services.Store.SDK": "10.1610.31001",
     "Newtonsoft.Json": "9.0.1",
     "NuSpec.ReferenceGenerator": "1.4.2"
